requirements.txt
llama-index
llama-index-readers-file
llama-index-vector-stores-postgres
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
llama-cpp-python
psycopg2-binary 
pgvector 
asyncpg 
sqlalchemy[asyncio] 
greenlet